Check npexp lims for sorted data before delete raw

+++ strategies.py
@@ -99,6 +99,16 @@ def delete_if_valid_backup_in_db(subject: dv.DataValidationFile, db: dv.DataVali
         if (subject.checksum != backups[0].checksum or subject.size != backups[0].size):
             raise AssertionError(f"Not a valid backup, something has gone wrong: {subject} {backups[0]}")
         
+        if (subject.path.startswith("A:") or subject.path.startswith("B:")) \
+            and subject.probe_dir and subject.probe_dir in ["ABC", "DEF"] \
+            and not (
+                (subject.npexp_path and any(s for s in subject.session.npexp_path.glob('*_sorted*')))
+                or (subject.lims_path and any(s for s in subject.session.lims_path.glob('*_sorted*')))
+            ):
+            # currently, we don't want to delete raw data on A/B drives before the sorted data make it to npexp
+            dv.logging.info(f"Skipped deletion of raw probe data: no sorted folders are on npexp or lims yet {subject.path} ")
+            return 0
+            
         try:
             pathlib.Path(subject.path).unlink()
             dv.logging.info(f"DELETED {subject.path}")
